Move navigation components to site folder and tidy up App.vue
