+++ Apps/WebClient/src/NewClientApp/src/components/site/RatingComponent.vue
