+++ Apps/WebClient/src/NewClientApp/src/components/site/SidebarComponent.vue
